feat: move settings data sections into components folders
Keep internal Settings -> Data section components out of route scanning by moving shared and stage-tamagotchi-only sections under components directories, and update imports accordingly.
