+++ apps/stage-tamagotchi/src/renderer/pages/settings/data/components/desktop-reset-section.vue
